GH-89 Include PID if application specifies a custom display name
An application can specify -Dvisualvm.display.name=some-name to specify a
custom name to be displayed in the "Applications" panel in VisualVM. If a
display name is specified, VisualVM doesn't include the PID in brackets after
the name.

This change causes VisualVM to include the PID when an application specifies a
custom name, which makes it easier to distinguish between different instances
of the same application.