Add ability to bulk select on FilteringToolbar
The filtering toolbar allows one to filter the selection. But there isn't any way to bulk select the view. This change adds that ability.

On top of that, it is an optional feature that subclasses can choose to implement or not.

The ThreadSelector is the single implementation allowing one to filter the threads then bulk toggle selection of those threads.